additional changes and refactoring

+++ tests/test_add_line_numbers.py
@@ -6,7 +6,7 @@ import os
 glb_new_line_symbol = "\n"
 glb_input_filename = "in_add_line_numbers.txt"
 glb_output_filename = "out_add_line_numbers.txt"
-
+param_remove_file_after_test = True
 
 def count_number_of_lines_in_file(a_file_name):
     counter = 0
@@ -32,11 +32,13 @@ class TestAddLineNumbers(unittest.TestCase):
 
     def tearDown(self):
         if os.path.exists(glb_input_filename):
-            os.remove(glb_input_filename)
+            if param_remove_file_after_test:
+                os.remove(glb_input_filename)
         else:
             print("can't find file", glb_input_filename, ". will not delete it.")
         if os.path.exists(glb_output_filename):
-            os.remove(glb_output_filename)
+            if param_remove_file_after_test:
+                os.remove(glb_output_filename)
         else:
             print("can't find file", glb_output_filename, ". will not delete it.")
