+++ tests/test_remove_empty_lines.py
@@ -4,6 +4,7 @@ import os
 
 glb_input_filename = "unit_test_input.txt"
 glb_output_filename = "unit_test_output.txt"
+param_remove_file_after_test = True
 
 
 def count_number_of_lines_in_file(a_file_name):
@@ -30,11 +31,13 @@ class TestRemoveEmptyLines(unittest.TestCase):
 
     def tearDown(self):
         if os.path.exists(glb_input_filename):
-            os.remove(glb_input_filename)
+            if param_remove_file_after_test:
+                os.remove(glb_input_filename)
         else:
             print("can't find file", glb_input_filename, ". will not delete it.")
         if os.path.exists(glb_output_filename):
-            os.remove(glb_output_filename)
+            if param_remove_file_after_test:
+                os.remove(glb_output_filename)
         else:
             print("can't find file", glb_output_filename, ". will not delete it.")
